requirements.txt
streamlit
requests
pandas
datetime
typing-extensions
streamlit_drawable_canvas
opencv-python
Pillow
